LayoutTests/accessibility/mac/option-with-lang-expected.txt


This tests that the lang attribute is returned appropriately for option elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectbox.childAtIndex(0).language is 'AXLanguage: ar'
PASS selectbox.childAtIndex(1).language is 'AXLanguage: en-US'
PASS successfullyParsed is true

TEST COMPLETE
